+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Improvements
 
+* Added `golang::from_tarball` to explicitly install Go from a binary tarball.
+  This can be used to make multiple installations as root or non-root users.
 * Use [`Stdlib::HTTPUrl`][] data type for URL parameters.
 
 [`Stdlib::HTTPUrl`]: https://github.com/puppetlabs/puppetlabs-stdlib/blob/0f032a9bc557949169f565bf41e5aa1f35b17346/REFERENCE.md#stdlibhttpurl