+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 * Added `golang::installation` to allow multiple installs on the same system.
   Installations can be owned by root or any other user.
+* Added `golang::linked_binaries` link binaries from a Go installation into a
+  `bin` directory.
 
 ### Improvements