+++ CHANGELOG.md
@@ -4,8 +4,17 @@ All notable changes to this project will be documented in this file.
 
 ## main branch
 
+### Features
+
+* Added `golang::installation` to allow multiple installs on the same system.
+  Installations can be owned by root or any other user.
+* Added `golang::linked_binaries` link binaries from a Go installation into a
+  `bin` directory.
+
 ### Improvements
 
+* Added `golang::from_tarball` to explicitly install Go from a binary tarball.
+  This can be used to make multiple installations as root or non-root users.
 * Use [`Stdlib::HTTPUrl`][] data type for URL parameters.
 
 [`Stdlib::HTTPUrl`]: https://github.com/puppetlabs/puppetlabs-stdlib/blob/0f032a9bc557949169f565bf41e5aa1f35b17346/REFERENCE.md#stdlibhttpurl