+++ CHANGELOG.md
@@ -23,9 +23,10 @@ All notable changes to this project will be documented in this file.
 
 ### Bug fixes
 
-* Updated minimum Puppet version to match puppet/archive. [archive version
-  4.0.0][https://forge.puppet.com/modules/puppet/archive/4.0.0] requires Puppet
-  5.5.8 or higher, so this module must as well.
+* Updated minimum Puppet version to match puppet/archive. [Archive version
+  4.0.0][archive4] requires Puppet 5.5.8 or higher, so this module must as well.
+
+[archive4]: https://forge.puppet.com/modules/puppet/archive/4.0.0
 
 ## Release 1.1.0